Question 1 and 2
I have gone through the code for plotting the graphs in question 1 and 2. I have found question 1 very tough, espicially getting through the data but i have got some plots at least. I am now going to move on to the code for the rest of the questions. I just wanted to get a commit in to make sure my code that runs is saved.