Code for Questions 4 and 5
This commit is to save the work I have done for question 4 and 5 so far. I am going to be looking at question 3 next, I have left this question for last. I feel like these questions were still tricky but better then the first question. Wish me luck for the rest of the paper!